+++ NEWS.md
@@ -1,3 +1,28 @@
+* Added additional tests for bug fixes and new features.
+    * Now up to 500 tests, all passing.
+    * Test coverage remains at 100%.
+* Added function `move_last_stage_to_useful()` and supporting functions
+  `form_C_mats()` and `form_eta_fu_phi_u_vecs()`.
+  These functions use a matrix method to move from last stage of final energy 
+  to last stage of useful energy.
+* Changed default number of rows in an allocation template from 3 to 4,
+  per request from Zeke Marshall. 
+  Note that the number of rows in the allocaiton template is adjustable 
+  at the time the template is created using the `n_allocation_rows` argument
+  to `fu_allocation_template()`.
+* Added functions for row and column notation in lists:
+  `switch_notation_byname()`, `arrow_to_paren_byname()`, and `paren_to_arrow_byname()`.
+  These functions are like `*_byname` functions in the package
+  `matsbyname`: they accept both single matrices and lists of matrices.
+  In that way, they are amenable to pipeline calculations using the 
+  `matsindf` package.
+  Adding this feature exposed bugs in `matsbyname::setrownames_byname()` and 
+  `matsbyname::setcolnames_byname()`. 
+  The bugs in `matsbyname` have been fixed, so these features are now working.
+* Added functions for row and column notation:
+  `paren_to_arrow()`, `arrow_to_paren()`, and `switch_notation()`.
+
+
 # IEATools 0.1.19 (2020-04-01)
 
 * New function `form_eta_fu_phi_u_vecs()` which